+++ frontend/src/components/Layout.tsx
@@ -1,5 +1,5 @@
 import { NavLink, Outlet, useNavigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext'
+import { useAuth } from '../contexts/useAuth'
 
 const NAV_ITEMS = [
   { to: '/dashboard', label: 'ダッシュボード', icon: '📊' },
